server/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
defusedxml==0.6.0
dj-rest-auth==2.1.2
Django==3.1.4
django-allauth==0.42.0
django-cors-headers==3.6.0
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycrypto==2.6.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
